initial commit
This is the main (experimental) functionality.  It works-ish on a basic dotorg site with the gutenberg plugin but needs lots of refinement, debugging and change.

It currently only addresses the setup where Gutenberg is loaded from a plugin and needs to be forced in a configured way to load.

It does not address the other case, where Gutenberg is in core and needs to be unloaded.

Much more to come -- at least we have a start.