Match the function name with the possible return value
Avoid confusion that can occur when `maybe_disable_gutenberg` returns `true` when gutenberg was going to be loaded because the class decision variable is called `load_gutenberg`.

If `load_gutenberg` is set to false (which reads as “don’t load”) - `maybe_disable_gutenberg` returns `false`, which reads as “don’t disable gutenberg”.

This aligns the method name with the expected result.