Refactor the way we're handling `load` values.
Change to remove any existing value for `load` when it is present in our new criteria instead of manipulating the value after the data is merged. This keeps `array_merge_recursive()` functioning as expected during the merge and prevents us from having to change the data we get back from it.