Fix array_key_exists warning
If no values are provided to the ramp function, the value is null, that was turning `$this->get(‘post_types’);` into false. When false is casted to an array - it’s casted as

```
[ false ]
```

And that produced unexpected results further up at `gutenberg_should_load` because `$criteria = $this->criteria->get();` returned an array with false inside.

Luckily, this didn’t produce a bug, but it did produce a warning - so that’s fixed now.